requirements.txt
openai-whisper
faster-whisper
pyannote.audio
feedparser
requests
python-dotenv
torch
torchaudio
customtkinter
openai
